Added easy to use makefiles to include in any project
These makefiles make the modplayer extensions easy to add to project with just with a single include statement in the project’s Makefile.

Because they figure out where they are relative to the parent makefile, they correctly modify the include paths and the correct path to the source files.

This relies on using the SDK’s common.mk for everything to work correctly.