Added setting to SYM_CPFLAGS
This variable is used in my custom common.mk which compiles simulator objects individually just like the target objects currently are. This allows passing compilation flags for the simulator objects.

If my mod to common.mk is accepted in the SDK then this will replace setting DYLIB_FLAGS in this case.